complete spa mall file

+++ lecture/nodejs_proficiency_week/remake_shopping_mall/SPA_MALL/routes/goods.js
@@ -1,6 +1,9 @@
 const express = require("express");
 const router = express.Router();
 
+const Goods = require("../schemas/goods.js");
+const Cart = require("../schemas/cart.js");
+const authMiddleware = require("../middlewares/auth-middleware.js");
 const goods = [
     {
         goodsId: 4,
@@ -47,10 +50,32 @@ router.get("/goods", async (req, res) => {
             price: item.price,
             thumbnailUrl: item.thumbnailUrl,
             category: item.category,
+            date: item.date,
         };
     });
 
     res.json({ goods: results });
+});
+// 장바구니 조회 API
+router.get("/goods/cart", authMiddleware, async (req, res) => {
+    const { userId } = res.locals.user;
+    console.log(userId);
+
+    const carts = await Cart.find({ userId }).exec();
+    const goodsIds = carts.map((cart) => cart.goodsId);
+
+    const goods = await Goods.find({ goodsId: goodsIds });
+
+    const results = carts.map((cart) => {
+        return {
+            quantity: cart.quantity,
+            goods: goods.find((item) => item.goodsId === cart.goodsId),
+        };
+    });
+
+    res.json({
+        carts: results,
+    });
 });
 
 //상품 상세 조회 API
@@ -67,18 +92,19 @@ router.get("/goods/:goodsId", async (req, res) => {
         price: goods.price,
         thumbnailUrl: goods.thumbnailUrl,
         category: goods.category,
+        date: goods.date,
     };
 
     res.json({ goods: result });
 });
 
-const Cart = require("../schemas/cart.js");
-
-router.post("/goods/:goodsId/cart", async (req, res) => {
+// 장바구니 등록 API
+router.post("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+    const { userId } = res.locals.user;
     const { goodsId } = req.params;
     const { quantity } = req.body;
 
-    const existsCarts = await Cart.find({ goodsId });
+    const existsCarts = await Cart.find({ userId, goodsId }).exec();
     if (existsCarts.length) {
         return res.status(400).json({
             success: false,
@@ -86,34 +112,37 @@ router.post("/goods/:goodsId/cart", async (req, res) => {
         });
     }
 
-    await Cart.create({ goodsId, quantity });
+    await Cart.create({ userId, goodsId, quantity });
 
     res.json({ result: "success" });
 });
 
-router.put("/goods/:goodsId/cart", async (req, res) => {
+// 장바구니 상품 수정 API
+router.put("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+    const { userId } = res.locals.user;
     const { goodsId } = req.params;
     const { quantity } = req.body;
 
-    const existsCarts = await Cart.find({ goodsId });
+    const existsCarts = await Cart.find({ userId, goodsId });
     if (existsCarts.length) {
-        await Cart.updateOne({ goodsId: goodsId }, { $set: { quantity: quantity } });
+        await Cart.updateOne({ userId, goodsId: goodsId }, { $set: { quantity: quantity } });
     }
     res.status(200).json({ success: true });
 });
 
-router.delete("/goods/:goodsId/cart", async (req, res) => {
+// 장바구니 상품 삭제 API
+router.delete("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+    const { userId } = res.locals.user;
     const { goodsId } = req.params;
 
-    const existsCarts = await Cart.find({ goodsId });
+    const existsCarts = await Cart.find({ userId, goodsId });
     if (existsCarts.length) {
-        await Cart.deleteOne({ goodsId });
+        await Cart.deleteOne({ userId, goodsId });
     }
 
     res.json({ result: "success" });
 });
 
-const Goods = require("../schemas/goods.js");
 router.post("/goods", async (req, res) => {
     const { goodsId, name, thumbnailUrl, category, price } = req.body;
